refactor(data): Include data file in data folder
