+++ components/navbar/counter-services.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { dataCounter } from "@/data";
+import { dataCounter } from "@/data/data";
 import CountUp from "react-countup";
 
 const CounterServices = () => {
